+++ gems/aws-sdk-s3/lib/aws-sdk-s3/plugins/md5s.rb
@@ -3,20 +3,13 @@ require 'openssl'
 module Aws
   module S3
     module Plugins
+      # @api private
       class Md5s < Seahorse::Client::Plugin
-
-        # Amazon S3 requires these operations to have an MD5 checksum
-        REQUIRED_OPERATIONS = [
-          :delete_objects,
-          :put_bucket_cors,
-          :put_bucket_lifecycle,
-          :put_bucket_lifecycle_configuration,
-          :put_bucket_policy,
-          :put_bucket_replication,
-          :put_bucket_tagging,
-          :put_object_legal_hold,
-          :put_object_lock_configuration,
-          :put_object_retention
+        # These operations allow Content MD5 but are not required by
+        # httpChecksumRequired
+        OPTIONAL_OPERATIONS = [
+          :put_object,
+          :upload_part
         ]
 
         # @api private
@@ -25,6 +18,7 @@ module Aws
           CHUNK_SIZE = 1 * 1024 * 1024 # one MB
 
           def call(context)
+            puts "------------i am an optional operation"
             body = context.http_request.body
             if body.size > 0
               context.http_request.headers['Content-Md5'] ||= md5(body)
@@ -64,19 +58,20 @@ module Aws
           doc_type: 'Boolean',
           docstring: <<-DOCS)
 When `true` a MD5 checksum will be computed for every request that
-sends a body.  When `false`, MD5 checksums will only be computed
-for operations that require them.  Checksum errors returned by Amazon
-S3 are automatically retried up to `:retry_limit` times.
+optionally allows for a Content MD5 header.  When `false`, MD5 checksums
+will not be computed for optional operations. Checksum errors returned by
+Amazon S3 are automatically retried up to `:retry_limit` times.
           DOCS
 
         def add_handlers(handlers, config)
-          # priority set low to ensure md5 is computed AFTER the request is
-          # built but before it is signed
-          handlers.add(Handler, {
-            priority: 10,
-            step: :build,
-            operations: config.compute_checksums ? nil : REQUIRED_OPERATIONS,
-          })
+          if config.compute_checksums
+            # priority set low to ensure md5 is computed AFTER the request is
+            # built but before it is signed
+            handlers.add(
+              Handler,
+              priority: 10, step: :build, operations: OPTIONAL_OPERATIONS
+            )
+          end
         end
 
       end