+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Support `httpChecksumRequired` on operations that require Content MD5 validation.
+
 3.97.1 (2020-06-01)
 ------------------